requirements.txt
streamlit
tensorflow-cpu
numpy
pillow
fpdf
matplotlib
gdown